% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertr-package.R
\docType{package}
\name{advertr-package}
\alias{advertr}
\alias{advertr-package}
\title{advertr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/advertr}
  \item Report bugs at \url{https://github.com/f-dallolio/advertr/issues}
}

}
\author{
\strong{Maintainer}: Filippo DallOlio \email{fdallolio@brocku.ca}

}
\keyword{internal}
