% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols.R
\name{get_cols_call}
\alias{get_cols_call}
\alias{get_cols_call.Field}
\alias{get_cols_call.Schema}
\alias{get_cols_call.Dataset}
\alias{get_cols_call.data.frame}
\title{Get column definitions}
\usage{
get_cols_call(x, to_str = TRUE)

\method{get_cols_call}{Field}(x, to_str = TRUE)

\method{get_cols_call}{Schema}(x, to_str = TRUE)

\method{get_cols_call}{Dataset}(x, to_str = TRUE)

\method{get_cols_call}{data.frame}(x, to_str = TRUE)
}
\arguments{
\item{x}{an objectg of class \code{Field}, \code{Schema}, \code{Dataset}, or \code{data.frame}.}

\item{to_str}{logical. If \code{TRUE} (default), \code{get_cols_call()} returns a string instead of a call.}
}
\value{
a list or a character vector.
}
\description{
Get column definitions
}
