% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols.R
\name{get_cols_type}
\alias{get_cols_type}
\alias{get_cols_type.Field}
\alias{get_cols_type.Schema}
\alias{get_cols_type.Dataset}
\alias{get_cols_type.data.frame}
\title{Get column definitions}
\usage{
get_cols_type(x)

\method{get_cols_type}{Field}(x)

\method{get_cols_type}{Schema}(x)

\method{get_cols_type}{Dataset}(x)

\method{get_cols_type}{data.frame}(x)
}
\arguments{
\item{x}{an object of class \code{Field}, \code{Schema}, \code{Dataset}, or \code{data.frame}.}
}
\value{
a list of Arrow data types.
}
\description{
Get column definitions
}
