% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-common.R
\name{path-common}
\alias{path-common}
\alias{path_common2}
\alias{path_common_remove}
\title{Detect or remove common path parts}
\usage{
path_common2(x, negate = FALSE, dir = FALSE)

path_common_remove(x, dir = FALSE)
}
\arguments{
\item{x}{a vector of paths.}

\item{negate}{logical. If \code{TRUE} the common part of the paths is removed. If \code{FALSE} (default), it is equivalent to \code{fs::patch_common()}.}

\item{dir}{logical. It applies \code{fs::path_dir()}.}
}
\value{
a vector of paths
}
\description{
Detect or remove common path parts
}
